UIMesh 좌표계 수정
-1~1 사이의 좌표계를
현재 킨 윈도우 크기에 맞게 수정

+++ Afx.cpp
@@ -1,7 +1,7 @@
 #include "Afx.h"
 
-int windowSize_W = 1600;
-int windowSize_H = 900;
+double windowSize_W = 1600;
+double windowSize_H = 900;
 
 double aspect_ratio = windowSize_W / windowSize_H;
